+++ client/src/App.js
@@ -1,5 +1,6 @@
 import './App.css';
 import {Switch, Route, Link} from 'react-router-dom';
+import {GoogleSignInButton} from './components/GoogleSignInButton';
 
 // components
 import RegisterServiceForm from './components/RegisterServiceForm';
@@ -13,7 +14,7 @@ function App() {
       <p>
         <Link to="/register-service">Register Service</Link>
       </p>
-
+      <GoogleSignInButton />
       <Switch>
         <Route path="/register-service">
           <RegisterServiceForm />