+++ client/src/components/GoogleSignInButton/index.js
@@ -1,0 +1,1 @@
+export {GoogleSignInButton} from './GoogleSignInButton';