+++ server/changelog.md
@@ -1,1 +1,7 @@
 <!-- add PR # and description -->
+
+### May 21
+- [] Add PassportJS, Google OAuth & express-session setup for social sign-in/sign-up
+- [] Update Mongoose models & fix model exports
+- [] Add setup/ability to use ES6 modules in server-side repo
+- [] Setup